Add check subscribe group

+++ src/handlers/callbacks.py
@@ -3,7 +3,7 @@ from aiogram.types import Message, CallbackQuery
 from sqlalchemy import select, insert
 
 from src.bot import bot
-from src.config import CHANNEL_NAME
+from src.config import CHANNEL_NAME, GROUP_NAME
 from src.db.models import User, Cart, Category, SubCategory, Product
 from src.db.base import get_session
 from src.keyboards import main_keyboard, generate_category_keyboard
@@ -42,9 +42,10 @@ async def get_products_handler(callback: CallbackQuery):
 @router.callback_query(F.data == 'check_subscribe')
 async def check_subscribe_handler(callback: CallbackQuery):
     user_channel_status = await bot.get_chat_member(chat_id=f'@{CHANNEL_NAME}', user_id=callback.from_user.id)
-    if user_channel_status.status != 'left':
+    user_group_status = await bot.get_chat_member(chat_id=f'@{GROUP_NAME}', user_id=callback.from_user.id)
+    if user_channel_status.status != 'left' and user_group_status.status != 'legt':
         await callback.answer(text='Спасибо за подписку!', show_alert=True)
         await callback.message.delete()
         await bot.send_message(callback.from_user.id, f'Привет, {callback.from_user.username}!', reply_markup=main_keyboard())
     else:
-        await callback.answer('Ты еще не подписался на канал', show_alert=True)
+        await callback.answer('Ты еще не подписался на канал или не вступил в группу', show_alert=True)
